Moved tests to different file
